+++ src/Symfony/Component/Cache/Tests/Adapter/AdapterTestCase.php
@@ -215,14 +215,9 @@ abstract class AdapterTestCase extends CachePoolTest
     }
 }
 
-class NotUnserializable implements \Serializable
+class NotUnserializable
 {
-    public function serialize()
-    {
-        return serialize(123);
-    }
-
-    public function unserialize($ser)
+    public function __wakeup()
     {
         throw new \Exception(__CLASS__);
     }